% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AppState}
\alias{AppState}
\title{R6 class representing a state of the back-end}
\description{
A state of the back-end
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AppState$new()}}
\item \href{#method-get_ncol}{\code{AppState$get_ncol()}}
\item \href{#method-clone}{\code{AppState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize default parameter(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppState$new(ncol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ncol}}{The number of columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ncol"></a>}}
\if{latex}{\out{\hypertarget{method-get_ncol}{}}}
\subsection{Method \code{get_ncol()}}{
Get the number of columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppState$get_ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
